feat(routes): Take away lofic from routes Move to user controllers
